+++ src/lib/djerba/extract/report_to_json.py
@@ -23,7 +23,7 @@ class composer_base(logger):
     # base class with shared methods and constants
 
     NA = 'NA'
-    ONCOGENIC = 'oncogenic'
+    ONCOGENIC = 'ONCOGENIC'
 
     def __init__(self, log_level=logging.WARNING, log_path=None):
         # list to determine sort order of oncokb level outputs
@@ -118,7 +118,16 @@ class clinical_report_json_composer(composer_base):
     V7_TARGET_SIZE = 37.285536 # inherited from CGI-Tools
     MSI_FILE = 'msi.txt'
 
-    EXCLUDED_VARIANT_CLASSIFICATIONS = ['Silent', 'Splice_Region']
+    # variant classifications excluded from TMB count
+    TMB_EXCLUDED = [
+        "3'Flank",
+        "3'UTR",
+        "5'Flank",
+        "5'UTR",
+        "Silent",
+        "Splice_Region",
+        "Targeted_Region",
+    ]
 
     UNCLASSIFIED_CYTOBANDS = [
         "", # some genes have an empty string for cytoband
@@ -158,10 +167,11 @@ class clinical_report_json_composer(composer_base):
         self.cytoband_map = self.read_cytoband_map()
         if self.failed:
             self.total_somatic_mutations = None
+            self.tmb_count = None
             self.total_fusion_genes = None
             self.gene_pair_fusions = None
         else:
-            self.total_somatic_mutations = self.read_total_somatic_mutations()
+            [self.total_somatic_mutations, self.tmb_count] = self.read_somatic_mutation_totals()
             if self.params.get(xc.ASSAY_TYPE) == rc.ASSAY_WGTS:
                 fus_reader = fusion_reader(input_dir, log_level=log_level, log_path=log_path)
                 self.total_fusion_genes = fus_reader.get_total_fusion_genes()
@@ -226,9 +236,9 @@ class clinical_report_json_composer(composer_base):
         # need to calculate TMB and percentiles
         cohort = self.read_cohort()
         data = {}
-        data[rc.TMB_TOTAL] = self.total_somatic_mutations
+        data[rc.TMB_TOTAL] = self.tmb_count
         # TODO See GCGI-347 for possible updates to V7_TARGET_SIZE
-        data[rc.TMB_PER_MB] = round(self.total_somatic_mutations/self.V7_TARGET_SIZE, 2)
+        data[rc.TMB_PER_MB] = round(self.tmb_count/self.V7_TARGET_SIZE, 2)
         data[rc.PERCENT_GENOME_ALTERED] = int(round(self.read_fga()*100, 0))
         csp = self.read_cancer_specific_percentile(data[rc.TMB_PER_MB], cohort, self.closest_tcga_lc)
         data[rc.CANCER_SPECIFIC_PERCENTILE] = csp
@@ -288,7 +298,7 @@ class clinical_report_json_composer(composer_base):
                     rc.VAF_PERCENT: int(round(float(input_row[self.TUMOUR_VAF]), 2)*100),
                     rc.TUMOUR_DEPTH: int(input_row[rc.TUMOUR_DEPTH]),
                     rc.TUMOUR_ALT_COUNT: int(input_row[rc.TUMOUR_ALT_COUNT]),
-                    rc.COPY_STATE: mutation_copy_states[gene],
+                    rc.COPY_STATE: mutation_copy_states.get(gene, self.UNKNOWN),
                     rc.ONCOKB: self.parse_oncokb_level(input_row)
                 }
                 rows.append(row)
@@ -572,19 +582,24 @@ class clinical_report_json_composer(composer_base):
         percentile = ecdf(tmb)*100
         return percentile
 
-    def read_total_somatic_mutations(self):
+    def read_somatic_mutation_totals(self):
+        # Count the somatic mutations
+        # Splice_Region is *excluded* for TMB, *included* in our mutation tables and counts
+        # Splice_Region mutations are of interest to us, but excluded from the standard TMB definition
+        # The TMB mutation count is (independently) implemented and used in vaf_plot.R
+        # See JIRA ticket GCGI-496
         total = 0
         excluded = 0
         with open(os.path.join(self.input_dir, self.MUTATIONS_EXTENDED)) as data_file:
             for row in csv.DictReader(data_file, delimiter="\t"):
                 total += 1
-                if row.get(self.VARIANT_CLASSIFICATION) in self.EXCLUDED_VARIANT_CLASSIFICATIONS:
+                if row.get(self.VARIANT_CLASSIFICATION) in self.TMB_EXCLUDED:
                     excluded += 1
-        count = total - excluded
-        msg = "Counted {} small mutations and indels; excluded {} of {}".format(count, excluded, total)+\
-              " based on variant classification"
+        tmb_count = total - excluded
+        msg = "Found {} small mutations and indels, of which {} are counted for TMB".format(total,
+                                                                                            tmb_count)
         self.logger.debug(msg)
-        return count
+        return [total, tmb_count]
 
     def build_genomic_biomarkers(self,input_dir,sample_ID):
         rows = []
@@ -771,8 +786,7 @@ class clinical_report_json_composer(composer_base):
         out_path = os.path.join(out_dir, 'msi.svg')
         args = [
             os.path.join(self.r_script_dir, 'biomarkers_plot.R'),
-            '-d', self.input_dir,
-            '-o', out_path
+            '-d', self.input_dir
         ]
         subprocess_runner(self.log_level, self.log_path).run(args)
         self.logger.info("Wrote biomarkers plot to {0}".format(out_path))
@@ -829,7 +843,7 @@ class fusion_reader(composer_base):
             fusion_genes.add(gene2)
             frame = fusion_data[fusion_id][0]['Frame']
             ann = annotations[fusion_id]
-            effect = ann['mutation_effect']
+            effect = ann['MUTATION_EFFECT']
             oncokb_level = self.parse_oncokb_level(ann)
             fda = self.parse_max_oncokb_level_and_therapies(ann, oncokb.FDA_APPROVED_LEVELS)
             [fda_level, fda_therapies] = fda